[Refactor] entity의 bigint 제거

+++ src/user-emoji/user-emoji.entity.ts
@@ -15,7 +15,7 @@ export class UserEmoji extends BaseTimeEntity {
   @PrimaryGeneratedColumn()
   id: number;
 
-  @Column({nullable:true, name: 'selected_order', type: 'bigint', default: null})
+  @Column({nullable:true, name: 'selected_order', type: 'int', default: null})
   selectedOrder: number;
 
   @ManyToOne(() => User, { eager: true })
